backend/requirements.txt
Flask==3.0.3
Pillow==10.2.0
torch==2.2.2
torchvision==0.17.2
facenet-pytorch==2.5.3
opencv-python==4.10.0.84
numpy==1.26.4
scikit-learn==1.4.2
ultralytics==8.3.38
torchreid==0.2.5
tqdm==4.66.2
